Split US file to match PSP, enable cod-esharing
